set display block for canvas element
Q: by default, canvas has display inline. And display inline can not set with and height. That's why it's recommended to set canvas as display block?

A: By default, a <canvas> element has a display type of inline, which means it behaves like a inline element, such as a span or an image.

One of the characteristics of inline elements is that they cannot have a width and height set explicitly, as they are designed to flow with the surrounding text.

However, a <canvas> element often needs to have a specific width and height set, in order to define the size of the drawing surface.

By setting the display type of the <canvas> element to block, you can overcome this limitation and set the width and height explicitly, using CSS properties like width and height.

So, setting display: block on a <canvas> element is a common practice, as it allows you to control the size of the canvas explicitly, which is often necessary for graphics and drawing applications.